Add files via upload
example of siimple doppler systems; these are often used to listen to fetal heart sounds or blood flow. They cannot be sterilized but surface cleanable. When needed in sterile settings like the operating room, the device can be placed in a sterile plastic sleeve.

Stethoscopes are never really sterile, so surface cleaning and/or use of a clean cover/sleeve is a reasonable approach when "contact precautions" are required for patient care.